requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Flask-Mail>=0.9.1
Werkzeug>=3.0.1
openpyxl>=3.1.2
pandas>=2.0.0
XlsxWriter>=3.1.9
reportlab>=4.0.0
# Hosting: required for Railway, Render, and most Python hosts
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
# Optional: only if using MySQL (DATABASE_URL=mysql+pymysql://...)
# PyMySQL>=1.1.0
